commands: refactor structure, tests and minimize imports
