+++ invenio_cli/commands/local.py
@@ -1,97 +1,28 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2019-2020 CERN.
-# Copyright (C) 2019-2020 Northwestern University.
+# Copyright (C) 2020 CERN.
 #
 # Invenio-Cli is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""Local and Containerized Commands."""
+"""Invenio module to ease the creation and management of applications."""
 
-import os
 import signal
-import subprocess
-from distutils import dir_util
-from pathlib import Path
+from distutils.dir_util import copy_tree
+from os import environ
+from pathlib import Path, PurePath
+from subprocess import PIPE, CalledProcessError
+from subprocess import Popen as popen
+from subprocess import run as run_proc
 
 import click
 from pynpm import NPMPackage
 
-from . import filesystem
-from .docker_helper import DockerHelper
-from .env import env
-from .services import HEALTHCHECKS, wait_for_services
-
-
-class Commands(object):
-    """Abstraction over CLI commands that are either local or containerized."""
-
-    def __init__(self, cli_config, docker_helper=None):
-        """Constructor.
-
-        :param cli_config: :class:CLIConfig instance
-        :param docker_helper: :class:DockerHelper instance, defaults to
-            containerized docker file (`docker-compose.full.yml`).
-        """
-        self.cli_config = cli_config
-        self.docker_helper = docker_helper or \
-            DockerHelper(cli_config.get_project_shortname(), local=False)
-
-    def shell(self):
-        """Start a shell in the virtual environment."""
-        command = ['pipenv', 'shell', ]
-        subprocess.run(command, check=True)
-
-    def pyshell(self, debug=False):
-        """Start a Python shell."""
-        with env(FLASK_ENV='development' if debug else 'production'):
-            command = ['pipenv', 'run', 'invenio', 'shell']
-            subprocess.run(command, check=True)
-
-    def status(self, services, verbose):
-        """Checks the status of the given service."""
-        project_shortname = self.cli_config.get_project_shortname()
-
-        for service in services:
-            check = HEALTHCHECKS.get(service)
-            if check:
-                if check(
-                    filepath="docker-services.yml",
-                    verbose=verbose,
-                    project_shortname=project_shortname,
-                ):
-                    click.secho(f"{service} up and running.", fg="green")
-                else:
-                    click.secho(
-                        f"{service}: unable to connect or bad status" +
-                        " response.",
-                        fg="red"
-                    )
-            else:
-                click.secho(
-                    f"{service}: no healthcheck function defined.",
-                    fg="yellow"
-                )
-
-    def stop(self):
-        """Stops containers."""
-        click.secho("Stopping containers...", fg="green")
-        self.docker_helper.stop_containers()
-        click.secho('Stopped containers', fg='green')
-
-    def destroy(self):
-        """Destroys the instance's virtualenv and containers."""
-        try:
-            subprocess.run(['pipenv', '--rm'], check=True)
-            click.secho('Virtual environment destroyed', fg='green')
-        except subprocess.CalledProcessError:
-            click.secho('The virtual environment was '
-                        'not removed as it was not '
-                        'created by pipenv', fg='red')
-
-        self.docker_helper.destroy_containers()
-        self.cli_config.update_services_setup(False)
-        click.secho('Destroyed containers', fg='green')
+from ..helpers import filesystem
+from ..helpers.docker_helper import DockerHelper
+from ..helpers.env import env
+from ..helpers.services import wait_for_services
+from .commands import Commands
 
 
 class LocalCommands(Commands):
@@ -112,7 +43,7 @@ class LocalCommands(Commands):
             command += ['--pre']
         if not lock:
             command += ['--skip-lock']
-        subprocess.run(command, check=True)
+        run_proc(command, check=True)
 
     def install_modules(self, modules):
         """Install modules."""
@@ -125,16 +56,16 @@ class LocalCommands(Commands):
             cmd.append(m)
 
         try:
-            subprocess.run(cmd, check=True)
-        except subprocess.CalledProcessError:
+            run_proc(cmd, check=True)
+        except CalledProcessError:
             click.secho('You must specify a valid path.', fg='red')
 
     def _update_instance_path(self):
         """Update path to instance in config."""
-        path = subprocess.run(
+        path = run_proc(
             ['pipenv', 'run', 'invenio', 'shell', '--no-term-title',
                 '-c', '"print(app.instance_path, end=\'\')"'],
-            check=True, universal_newlines=True, stdout=subprocess.PIPE
+            check=True, universal_newlines=True, stdout=PIPE
         ).stdout.strip()
         self.cli_config.update_instance_path(path)
 
@@ -150,11 +81,12 @@ class LocalCommands(Commands):
         assets = 'assets'
         click.secho('Symlinking {}...'.format(assets), fg='green')
 
-        instance_path = str(self.cli_config.get_instance_path()) + '/'
-        project_dir = str(self.cli_config.get_project_dir())
+        instance_path = self.cli_config.get_instance_path()
+        project_dir = self.cli_config.get_project_dir()
         for file_path in files_to_link:
-            relpath = file_path.split(instance_path)[1]
-            target_path = os.path.join(project_dir, relpath)
+            file_path = Path(file_path)
+            relative_path = file_path.relative_to(instance_path)
+            target_path = project_dir / relative_path
             filesystem.force_symlink(target_path, file_path)
 
     def _copy_statics_and_assets(self):
@@ -167,7 +99,7 @@ class LocalCommands(Commands):
         dst_dir = self.cli_config.get_instance_path() / static
         dst_dir = str(dst_dir)
         # using it for a different purpose then intended but very useful
-        dir_util.copy_tree(src_dir, dst_dir)
+        copy_tree(src_dir, dst_dir)
 
         assets = 'assets'
         src_dir = self.cli_config.get_project_dir() / assets
@@ -175,7 +107,7 @@ class LocalCommands(Commands):
         dst_dir = self.cli_config.get_instance_path() / assets
         dst_dir = str(dst_dir)
         # The full path to the files that were copied is returned
-        copied_files = dir_util.copy_tree(src_dir, dst_dir)
+        copied_files = copy_tree(src_dir, dst_dir)
         return copied_files
 
     def update_statics_and_assets(self, force, flask_env='production'):
@@ -191,15 +123,15 @@ class LocalCommands(Commands):
         with env(FLASK_ENV=flask_env):
             # Collect into statics/ and assets/ folder
             click.secho('Collecting statics and assets...', fg='green')
-            subprocess.run(collect_cmd, check=True)
+            run_proc(collect_cmd, check=True)
             if force:
-                subprocess.run(clean_create_cmd, check=True)
+                run_proc(clean_create_cmd, check=True)
 
                 # Installs in assets/node_modules/
                 click.secho('Installing JS dependencies...', fg='green')
-                subprocess.run(install_cmd, check=True)
+                run_proc(install_cmd, check=True)
             else:
-                subprocess.run(create_cmd, check=True)
+                run_proc(create_cmd, check=True)
 
             # Symlink the instance's statics and assets
             copied_files = self._copy_statics_and_assets()
@@ -207,7 +139,7 @@ class LocalCommands(Commands):
 
             # Build project
             click.secho('Building assets...', fg='green')
-            subprocess.run(build_cmd, check=True)
+            run_proc(build_cmd, check=True)
 
     def watch_assets(self):
         """High-level command to watch assets for changes."""
@@ -219,7 +151,7 @@ class LocalCommands(Commands):
             # Collect into statics/ and assets/ folder
             click.secho('Starting assets watching (press CTRL+C to stop)...',
                         fg='green')
-            subprocess.run(watch_cmd, check=True)
+            run_proc(watch_cmd, check=True)
 
     @staticmethod
     def _watch_js_module(pkg):
@@ -295,13 +227,13 @@ class LocalCommands(Commands):
                 'pipenv', 'run', 'invenio', 'shell', '--no-term-title', '-c',
                 "import redis; redis.StrictRedis.from_url(app.config['CACHE_REDIS_URL']).flushall(); print('Cache cleared')"  # noqa
             ]
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             # TODO: invenio-db#126 should make it idempotent
             command = [
                 'pipenv', 'run', 'invenio', 'db', 'destroy', '--yes-i-know',
             ]
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             # TODO: invenio-indexer#114 should make destroy and queue init
             #       purge idempotent
@@ -309,42 +241,42 @@ class LocalCommands(Commands):
                 'pipenv', 'run', 'invenio', 'index', 'destroy',
                 '--force', '--yes-i-know',
             ]
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             command = [
                 'pipenv', 'run', 'invenio', 'index', 'queue', 'init', 'purge']
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             self.cli_config.update_services_setup(False)
 
         if not self.cli_config.get_services_setup():
             command = ['pipenv', 'run', 'invenio', 'db', 'init', 'create']
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             command = [
                 'pipenv', 'run', 'invenio', 'files', 'location', 'create',
                 '--default', 'default-location',
                 "{}/data".format(self.cli_config.get_instance_path())
             ]
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             # Without the self.cli_config.get_services_setup() check
             # this throws an error on re-runs
             # TODO: invenio-accounts#297 should make it idempotent
             command = ['pipenv', 'run', 'invenio', 'roles', 'create', 'admin']
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             command = [
                 'pipenv', 'run', 'invenio', 'access', 'allow',
                 'superuser-access', 'role', 'admin'
             ]
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             # Without the self.cli_config.get_services_setup() check
             # this throws an error on re-runs
             # TODO: invenio-indexer#115 should make it idempotent
             command = ['pipenv', 'run', 'invenio', 'index', 'init']
-            subprocess.run(command, check=True)
+            run_proc(command, check=True)
 
             self.cli_config.update_services_setup(True)
 
@@ -353,7 +285,7 @@ class LocalCommands(Commands):
         self._ensure_containers_running()
 
         command = ['pipenv', 'run', 'invenio', 'rdm-records', 'demo']
-        subprocess.run(command, check=True)
+        run_proc(command, check=True)
 
     def run(self, host, port, debug):
         """Run development server and celery queue."""
@@ -368,15 +300,15 @@ class LocalCommands(Commands):
         signal.signal(signal.SIGINT, signal_handler)
 
         click.secho("Starting celery worker...", fg="green")
-        worker = subprocess.Popen([
+        worker = popen([
             'pipenv', 'run', 'celery', '--app', 'invenio_app.celery', 'worker'
         ])
 
         click.secho("Starting up local (development) server...", fg='green')
-        run_env = os.environ.copy()
+        run_env = environ.copy()
         run_env['FLASK_ENV'] = 'development' if debug else 'production'
         run_env['INVENIO_SITE_HOSTNAME'] = f"{host}:{port}"
-        server = subprocess.Popen([
+        server = popen([
             'pipenv', 'run', 'invenio', 'run', '--cert',
             'docker/nginx/test.crt', '--key', 'docker/nginx/test.key',
             '--host', host, '--port', port
@@ -386,139 +318,3 @@ class LocalCommands(Commands):
             'Instance running!\nVisit https://{}:{}'.format(host, port),
             fg='green')
         server.wait()
-
-
-class ContainerizedCommands(Commands):
-    """Containerized environment CLI commands."""
-
-    def __init__(self, cli_config, docker_helper=None):
-        """Constructor."""
-        docker_helper = docker_helper or \
-            DockerHelper(cli_config.get_project_shortname(), local=False)
-
-        super(ContainerizedCommands, self).__init__(cli_config, docker_helper)
-
-    def update_statics_and_assets(self, install):
-        """Update application containers' statics and assets."""
-        project_shortname = self.cli_config.get_project_shortname()
-
-        click.secho('Collecting statics and assets...', fg='green')
-        # Collect into statics/ folder
-        self.docker_helper.execute_cli_command(
-            project_shortname, 'invenio collect')
-
-        # Collect into assets/ folder
-        self.docker_helper.execute_cli_command(
-            project_shortname, 'invenio webpack create')
-
-        if install:
-            click.secho('Installing js dependencies...', fg='green')
-            # Installs in assets/node_modules/
-            self.docker_helper.execute_cli_command(
-                project_shortname, 'invenio webpack install --unsafe')
-
-        # Use the copy approach rather than the symlink one in container
-        click.secho('Copying project statics and assets...', fg='green')
-        src_dir = self.cli_config.get_project_dir() / 'static'
-        # src_dir MUST be '<dir>/.' for docker cp to work
-        src_dir = os.path.join(src_dir, '.')
-        dst_dir = '/opt/invenio/var/instance/static/'
-        self.docker_helper.copy2(src_dir, dst_dir)
-
-        src_dir = self.cli_config.get_project_dir() / 'assets'
-        # src_dir MUST be '<dir>/.' for docker cp to work
-        src_dir = os.path.join(src_dir, '.')
-        dst_dir = '/opt/invenio/var/instance/assets/'
-        self.docker_helper.copy2(src_dir, dst_dir)
-
-        click.secho('Building assets...', fg='green')
-        self.docker_helper.execute_cli_command(
-            project_shortname, 'invenio webpack build')
-
-    def _lock_python_dependencies(self, pre):
-        """Creates Pipfile.lock if not existing."""
-        command = ['pipenv', 'lock']
-
-        if pre:
-            command += ['--pre']
-
-        locked = 'Pipfile.lock' in os.listdir('.')
-        if not locked:
-            subprocess.run(command, check=True)
-
-    def containerize(self, pre, force, install):
-        """Launch fully containerized application."""
-        self._lock_python_dependencies(pre)
-
-        click.secho('Making sure containers are up...', fg='green')
-        self.docker_helper.start_containers()
-
-        project_shortname = self.cli_config.get_project_shortname()
-
-        wait_for_services(
-            services=["redis", self.cli_config.get_db_type(), "es"],
-            project_shortname=project_shortname,
-        )
-
-        if force:
-            self.cli_config.update_services_setup(False)
-            click.secho("Flushing redis cache...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname,
-                "invenio shell --no-term-title -c "
-                "\"import redis; redis.StrictRedis.from_url(app.config['CACHE_REDIS_URL']).flushall(); print('Cache cleared')\""  # noqa
-            )
-            click.secho("Deleting database...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname, 'invenio db destroy --yes-i-know')
-            click.secho("Deleting indices...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname,
-                'invenio index destroy --force --yes-i-know')
-            click.secho("Purging queues...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname, 'invenio index queue init purge')
-
-        if not self.cli_config.get_services_setup():
-            click.secho("Creating database...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname, 'invenio db init create')
-
-            click.secho("Creating files location...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname,
-                "invenio files location create --default default-location "
-                "${INVENIO_INSTANCE_PATH}/data"
-            )
-
-            click.secho("Creating admin role...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname, 'invenio roles create admin')
-
-            click.secho(
-                "Assigning superuser access to admin role...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname,
-                'invenio access allow superuser-access role admin')
-
-            click.secho("Creating indices...", fg="green")
-            self.docker_helper.execute_cli_command(
-                project_shortname, 'invenio index init')
-
-            self.cli_config.update_services_setup(True)
-
-        # statics and assets are always regenerated
-        self.update_statics_and_assets(install=install)
-
-        click.secho(
-            'Instance running!\nVisit https://127.0.0.1', fg='green')
-
-    def demo(self):
-        """Add demo records into the instance."""
-        project_shortname = self.cli_config.get_project_shortname()
-
-        # TODO: Shall we call ensure containers? If so move to
-        # parent object. Noentheless, the way of starting in
-        # container mode ensures the containers are up
-        self.docker_helper.execute_cli_command(
-            project_shortname, 'invenio rdm-records demo')
