+++ invenio_cli/helpers/filesystem.py
@@ -11,7 +11,7 @@
 
 import errno
 import hashlib
-import os
+from os import listdir, remove, symlink
 from os.path import isdir
 from pathlib import Path
 
@@ -36,7 +36,7 @@ def hash_file(path_to_file):
 def get_created_files(folder):
     """Return the generated tree of files (and their hash) and folders."""
     files = {}
-    for name in os.listdir(folder):
+    for name in listdir(folder):
         path = Path(folder)  # Current path
 
         # Add files and their hash
@@ -52,8 +52,8 @@ def get_created_files(folder):
 def force_symlink(target, link_name):
     """Forcefully create symlink at link_name pointing to target."""
     try:
-        os.symlink(target, link_name)
+        symlink(target, link_name)
     except OSError as e:
         if e.errno == errno.EEXIST:
-            os.remove(link_name)
-            os.symlink(target, link_name)
+            remove(link_name)
+            symlink(target, link_name)
