+++ tests/commands/testpkg/package.json
